refactor: rename and extract createPlanning components
