+++ src/features/personal/createPlanning/form/CreatePlanningSlides/AerobicInput/AerobicInput.module.scss
