+++ src/features/personal/createPlanning/form/CreatePlanningSlides/CreateTrainingSwiper/CreateTrainingSwiper.tsx
