+++ src/features/personal/createPlanning/form/CreatePlanningSlides/GoToTrainingCheckoutButton/GoToTrainingCheckoutButton.tsx
