+++ src/features/personal/createPlanning/form/CreatePlanningSlides/MuscleGroupsPreview/index.tsx
