+++ src/features/personal/createPlanning/form/CreatePlanningSlides/MuscleGroupsPreview/styles.module.scss
