+++ src/features/personal/createPlanning/form/CreatePlanningSlides/SetSlide/ExerciseTechniquesRadios/ExerciseTechniqueRadio.tsx
