+++ src/features/personal/createPlanning/form/CreatePlanningSlides/SetSlide/ExercisesSelects/CustomSelect.tsx
