+++ src/features/personal/createPlanning/form/CreatePlanningSlides/SetSlide/ExercisesSelects/EquipmentsSelect/index.tsx
