+++ src/features/personal/createPlanning/form/CreatePlanningSlides/SetSlide/ExercisesSelects/ExercisesSelects.tsx
