+++ src/features/personal/createPlanning/form/CreatePlanningSlides/SetSlide/ExercisesSelects/GripsSelect/index.tsx
