+++ src/features/personal/createPlanning/form/CreatePlanningSlides/TrainingCheckout/TrainingCheckout.module.scss
