+++ src/features/personal/createPlanning/form/CreatePlanningSlides/TrainingCheckout/TrainingCheckout.tsx
